fixed autostart, hopefully hyprland will autostart and yafti will run as intented

+++ usr/share/ublue-os/firstboot/launcher/autostart.sh
@@ -1,11 +1,17 @@
 #!/usr/bin/env bash
 
-cp -r /usr/etc/homedir/* ~HOME/
-cp -r /usr/etc/homedir/.config/ ~/.config/
-cp -r /usr/etc/homedir/.local/share/ ~/.local/share/
+# Ensure target file exists and is a symlink (not a regular file or dir).
+if [ ! -f "$HOME"/.config/ublue-firstboot ]; then
+    # copy config files
+    cp -r /usr/etc/homedir/* ~HOME/
+    cp -r /usr/etc/homedir/.config/ ~/.config/
+    cp -r /usr/etc/homedir/.local/share/ ~/.local/share/
 
-Hyprland
+    # add file to prevent firstboot from running again
 
-sleep 10
-# Simply launches the "yafti" GUI with the uBlue image's configuration.
-/usr/bin/yafti /usr/share/ublue-os/firstboot/yafti.yml
+    touch "$HOME"/.config/ublue-firstboot
+
+    sed -i " 1 d" "$HOME"/.config/hypr/hyprland.conf
+
+    /usr/bin/yafti /usr/share/ublue-os/firstboot/yafti.yml
+fi
